XFCE4 settings are started automatically, clipman as well. And Mice splash also.

+++ xfce_settings.sh
@@ -1,43 +1,44 @@
 #!/usr/bin/env bash
-# Additional changes to XFCE settings (xsettings.xml)
-xfconf-query -c xsettings -p /Net/ThemeName -s "Breeze-Dark"
-xfconf-query -c xsettings -p /Net/IconThemeName -s "Papirus-Dark"
-xfconf-query -c xsettings -p /Gtk/FontName -s "Ubuntu 10"
+# Change some settings of XFCE4 look & feel
 
-# Change window style manager
-xfconf-query -c xfwm4 -p /general/theme -s "Moheli" # TODO: Mint-y-dark-blue
-xfconf-query -c xfwm4 -p /general/title_font -s "Ubuntu Medium 10"
+if [ ! -f /home/$USER/.local/.xfce_settings_done ]; then
+  # Additional changes to XFCE settings (xsettings.xml)
+  xfconf-query -c xsettings -p /Net/ThemeName -s "Breeze-Dark"
+  xfconf-query -c xsettings -p /Net/IconThemeName -s "Papirus-Dark"
+  xfconf-query -c xsettings -p /Gtk/FontName -s "Ubuntu 10"
 
-# Move terminalrc config to correct location
-mkdir -p /home/$USER/.config/xfce4/terminal/
-cp -rf /app/terminalrc /home/$USER/.config/xfce4/terminal/
+  # Change window style manager
+  # TODO: Change theme, like: Mint-y-dark-blue
+  xfconf-query -c xfwm4 -p /general/theme -s "Moheli"
+  xfconf-query -c xfwm4 -p /general/title_font -s "Ubuntu Medium 10"
 
-# Use custom whisker menu settings
-mkdir -p /home/$USER/.config/xfce4/panel/
-cp -rf /app/whiskermenu-1.rc /home/$USER/.config/xfce4/panel/
+  # Move terminalrc config to correct location
+  mkdir -p /home/$USER/.config/xfce4/terminal/
+  cp -rf /app/terminalrc /home/$USER/.config/xfce4/terminal/
 
-# Change default menu to whisker menu
-xfconf-query -c xfce4-panel -p /plugins/plugin-1 -s "whiskermenu"
+  # Use custom whisker menu settings
+  mkdir -p /home/$USER/.config/xfce4/panel/
+  cp -rf /app/whiskermenu-1.rc /home/$USER/.config/xfce4/panel/
 
-# Change bottom-panel to intelligently-hide
-xfconf-query -c xfce4-panel -p /panels/panel-2/autohide-behavior -n -t int -s 1
+  # Change default menu to whisker menu
+  xfconf-query -c xfce4-panel -p /plugins/plugin-1 -s "whiskermenu"
 
-# Change Splash engine
-xfconf-query -c xfce4-session -p /splash/Engine -n -t string -s "mice"
+  # Change bottom-panel to intelligently-hide
+  xfconf-query -c xfce4-panel -p /panels/panel-2/autohide-behavior -n -t int -s 1
 
-# Reset panel(s)
-xfce4-panel -r
+  # Maybe disable compositor (if not already), is that wise?
+  # TODO: Add PulseAudio plugin to panel
 
-# Be sure XFCE4 session is restarted fully
-sleep 1
+  # Panels are getting created wait before xfce setup is completed
+  sleep 1
 
-# TODO: Use Mice Splash (session and startup)
-# TODO: Use whiskermenu menu as default menu (add plugin to panel) 
-# TODO: Maybe disable compositor (if not already)?
-# TODO: Add PulseAudio plugin to panel
-# TODO: Remove screen locker from autostart and check Clipman to actually autostart
+  # Change browser icon to Firefox
+  launchName="launcher-11"
+  filename=$(ls /home/$USER/.config/xfce4/panel/$launchName/ | head -1)
+  sed -i "s/Icon=.*/Icon=firefox/" /home/$USER/.config/xfce4/panel/$launchName/$filename
 
-# Change browser icon to Firefox
-launchName="launcher-11"
-filename=$(ls /home/$USER/.config/xfce4/panel/$launchName/ | head -1)
-sed -i "s/Icon=.*/Icon=firefox/" /home/$USER/.config/xfce4/panel/$launchName/$filename
+  # Reset panel(s) once
+  xfce4-panel -r
+
+  touch /home/$USER/.local/.xfce_settings_done
+fi
